Introducing sample code snippets
This change introduce some sample code snippet to illustrate the
usage of the library on its core features. For now only a few use
cases are demonstrated but it is just the beginning.

The drawback here is that I had to reorganize all the files to
separate everything into gradle subprojects...