Ability to Refresh the user cache on the session
This seemed necessary. If new users are added post connection by the
session I do not want to have to disconnect and reconnect the session to
pull the users in again. I simply and blindly replaced the user cache to
the current status of the slack instance state. If you would like to
maintain users across refresh of the user cache update (say a user is
deleted in between) you should store it away in persistent storage
outside of the session.