FIX not throwing anymore on unknown event
Prior to this change, an IllegalArgumentException was thrown when
receiving an unknown event. This fix replaces the exception by a
warning in the logs. There's no point to throw an exception in this
case, at best on this kind of event, nothing has to be done and since
the Slack RTM API is evolving more and more events not yet supported
can appears.